playwright: factor out helper for setting up 4S
We seem to already have about 5 copies of this code, so before I add another,
let's factor it out.
